Fix audio balance below 50%

+++ bsnes/target-bsnes/program/audio.cpp
@@ -74,6 +74,6 @@ auto Program::updateAudioEffects() -> void {
   double volume = settings.audio.volume * 0.01;
   Emulator::audio.setVolume(volume);
 
-  double balance = max(-1.0, min(+1.0, (settings.audio.balance - 50) / 50.0));
+  double balance = max(-1.0, min(+1.0, ((int)settings.audio.balance - 50) / 50.0));
   Emulator::audio.setBalance(balance);
 }
